claude: move one_of/optional schema-shape tests into integration tests
These tests didn't actually need access to private items — they only inspect
public schema/parse_raw output — so embedding them in src/generators/combinators.rs
forced us to duplicate schema_type / one_of_children helpers (one copy in
tests/common/utils.rs for the integration tests, another inline in the embedded
test file). Moving them next to the rest of test_combinators.rs lets them
reuse the existing helpers and drops the duplication.
