+++ tests/test_combinators.rs
@@ -1,7 +1,8 @@
 mod common;
 
-use common::utils::find_any;
+use common::utils::{find_any, one_of_children, schema_type};
 use hegel::TestCase;
+use hegel::ciborium::Value;
 use hegel::generators::{self as gs, Generator};
 
 #[hegel::test]
@@ -178,4 +179,68 @@ fn test_optional_mapped_find_any() {
         gs::optional(gs::integers::<i32>().map(|n| n.wrapping_mul(2))),
         |v| v.is_none(),
     );
+}
+
+#[test]
+fn test_one_of_schema_is_flat_one_of() {
+    let g = gs::one_of(vec![
+        gs::booleans().boxed(),
+        gs::booleans().boxed(),
+        gs::booleans().boxed(),
+    ]);
+    let basic = g.as_basic().expect("one_of of basics should be basic");
+    let schema = basic.schema();
+
+    assert_eq!(schema_type(schema), "one_of");
+
+    // Children are emitted directly with no tagged-tuple wrapping.
+    let children = one_of_children(schema);
+    assert_eq!(children.len(), 3);
+    for child in children {
+        assert_eq!(schema_type(child), "boolean");
+    }
+}
+
+#[test]
+fn test_one_of_basic_dispatches_by_index() {
+    // Two children with map transforms; the wire response selects which.
+    let g = gs::one_of(vec![
+        gs::booleans().map(|_| "first".to_string()).boxed(),
+        gs::booleans().map(|_| "second".to_string()).boxed(),
+    ]);
+    let basic = g.as_basic().expect("one_of of basics should be basic");
+
+    // Simulate the server response: [index, value].
+    let raw_first = Value::Array(vec![Value::Integer(0.into()), Value::Bool(false)]);
+    let raw_second = Value::Array(vec![Value::Integer(1.into()), Value::Bool(true)]);
+
+    assert_eq!(basic.parse_raw(raw_first), "first");
+    assert_eq!(basic.parse_raw(raw_second), "second");
+}
+
+#[test]
+fn test_optional_schema_is_flat_one_of() {
+    let g = gs::optional(gs::booleans());
+    let basic = g.as_basic().unwrap();
+    let schema = basic.schema();
+
+    assert_eq!(schema_type(schema), "one_of");
+
+    let children = one_of_children(schema);
+    assert_eq!(children.len(), 2);
+    // First child is `null`, second is the inner schema (`boolean`).
+    assert_eq!(schema_type(&children[0]), "null");
+    assert_eq!(schema_type(&children[1]), "boolean");
+}
+
+#[test]
+fn test_optional_basic_dispatches_by_index() {
+    let g = gs::optional(gs::booleans());
+    let basic = g.as_basic().unwrap();
+
+    let raw_none = Value::Array(vec![Value::Integer(0.into()), Value::Null]);
+    let raw_some = Value::Array(vec![Value::Integer(1.into()), Value::Bool(true)]);
+
+    assert_eq!(basic.parse_raw(raw_none), None);
+    assert_eq!(basic.parse_raw(raw_some), Some(true));
 }
